prometheus-nats-adapter/requirements.txt
flask[async]
protobuf==3.20.0
nats-py
requests
python-snappy
